% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_html.R
\name{fetch_datalist_urls}
\alias{fetch_datalist_urls}
\title{Fetch URLs Of Description Pages About Each Identifiers}
\usage{
fetch_datalist_urls()
}
\description{
Fetch URLs Of Description Pages About Each Identifiers
}
