% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_urls.R
\name{fetch_description_urls}
\alias{fetch_description_urls}
\title{Fetch URLs Of Description Pages About Each Identifiers}
\usage{
fetch_description_urls()
}
\description{
Fetch URLs Of Description Pages About Each Identifiers
}
