% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kokudosuuchiUtils-package.R
\docType{package}
\name{kokudosuuchiUtils}
\alias{kokudosuuchiUtils}
\alias{kokudosuuchiUtils-package}
\title{Utilities For 'kokudosuuchi' Package}
\description{
Utilities For 'kokudosuuchi' Package
}
